Automatic reporting to OPEN-CUTS (#1415)
WOOT WOOT, IT'S AUTOMATIC REPORTING! With [OPEN-CUTS](https://ubports.open-cuts.org) coming along nicely, we can begin looking at experimental automatic reporting! This resolves #868. By specifying the env vars `OPENCUTS=true` (for anonymous reporting) or `OPENCUTS_API_KEY=<token linked to your account>` (for reporting associated with your account), the installer will:

- automatically report a `PASS` run to OPEN-CUTS, once an installation has finished successfully;
- automatically report a `WONKY` run to OPEN-CUTS and attach it to a user-reqested bug report;
- automatically report a `FAIL` bug report to OPEN-CUTS and attach it to a bug report that originated in an unhandled error.

The OPEN-CUTS runs are created with a log, the combination metadata, and a comment describing the situation. In the future, we should add more UI to ask users for their results and eliminate the need for the env vars, but for now it's a good solution, as these features are still experimental.

This PR also cleans up the reporting file and prepares it for potential future modularization. JSdoc comments have been added to all functions in that file (part of #956). The generated issue template has been tweaked to improve access to essential information. The bug report will now include the full error message (resolves #693) and the title will be shortened in the middle if it is too long. Previously, the issue template was sometimes not properly URL-encoded, leading the issue to be cut off halfway. This has been fixed now.

* First crack at automatic run creation

* Lint

* Rename report library

* WIP: Clean up reporting

* Automatically report successful runs (PASS) to OPEN-CUTS if OPENCUTS or OPEN_CUTS_KEY environment variable is set. Resolves #868. Automatically report a WONKY run on a user-requested bug report and a FAIL run on an unhandled error and add the link to the bugreport. Clean up reporting library and add JSdoc comments (part of #956). Include full error in bugreport (resolves #693) and ensure a maximum length for issue titles and ensure proper URI encoding.

* Remove long paths from error title