Electron 10 babyyy (#1425)
Finally get rid of dependabot's annoying warning, this is the latest version of electron now. Jumping five major releases in one go was kind of a pain in the ass, because some libraries went unmaintained in the mean time and did not keep up with the api changes, but i think everything should work now.