plugins: manual_download: Resolve copyFile promise
The whole block beginning with `.then(downloadedFilePath`
and ending with `copyFile()` did not return anything and was
never going to resolve its promise with an actual return
value.
That meant that that `checkFile()` was being called before
any of the promise functions inside the previous block had
the chance to run, therefore the copied file was not yet
available, leading to checkFile trying to verify a
non-existent path, leading to a checksum mismatch error.

Solution is to make the whole block return something, in
this case the value of the last chained promise, by removing
the curly braces.

Solves:
```
error: Error: core:manual_download: Error: checksum mismatch
```

Fixes https://github.com/ubports/ubports-installer/issues/1754