Handle Heimdall missing MSVC++2012
If Heimdall is started on Windows without the Microsoft Visual C++ 2012
Redistributable package installed, it will exit with status 3221225781.

Since we're treating errors in wait() as fatal as of 3fa043cf (and
reinforced that decision in 4f3899cd), catch and notify the user of this
situation. They can choose to ignore it if they aren't installing on a
Samsung device.

I had to make some changes to the promise chains in the plugin index
and in core.js. The changes to the plugin index caused several unit
tests which were relying on very exact return values to fail, so they
had to be refactored.

Some tests didn't accurately describe or test their stated goals, so I
fixed them up.

This improves logging quite a lot.