Disable manual download buttons during checksum test (#2622)
After manually downloading a file, a checksum test is performed on the
file. For large files this test can take some time (seconds), and
there's currently no indication in the UI that this is taking place.

This change disables the file selector and "Continue" buttons during the
checksum test, changing the text in the button to "Checking..." instead.
This provides immediate feedback for the user and avoids the confusion
of allowing them to still use the buttons while the check is taking
place.