plugins: lineage: select latest build instead of oldest
The response is sorted from newest to oldest version.
We are selecting the last entry (length - 1), which is the oldest
version.

This results in devices with support for multiple lineage versions
(eg 19.1, 20.0) to retrieve the older version (eg 19.1) which
is in most cases unsupported.

If the device only supports a single version, it selects older
builds as well, which will ask to update again.

Use the first entry of the response to install the latest version.

Fixes: https://github.com/ubports/ubports-installer/issues/3141
Change-Id: I7f37abfc29f03da45e85e1fe5f4cc769282f953f
Signed-off-by: Alexander Martinz <amartinz@shiftphones.com>